Safari-609.1.11.1/LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/browsers/history/the-location-interface/assign_after_load-expected.txt


FAIL Assignment to location after document is completely loaded assert_equals: expected 3 but got 2
